+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### Changed
 
-- nothing changed
+- fixed password change token expiry computation
+- fixed register serializer to allow null values of url_format and email_format
 
 ### Removed